+++ src/Alley/Validator/ContainsString.php
@@ -14,33 +14,54 @@ declare(strict_types=1);
 namespace Alley\Validator;
 
 use Laminas\Validator\Exception\InvalidArgumentException;
-use Laminas\Validator\IsInstanceOf;
-use Laminas\Validator\ValidatorInterface;
 
 final class ContainsString extends ExtendedAbstractValidator
 {
     public const NOT_CONTAINS_STRING = 'notContainsString';
 
-    protected $messageTemplates = [
+    protected array $messageTemplates = [
         self::NOT_CONTAINS_STRING => 'Must contain string "%needle%".',
     ];
 
-    protected $messageVariables = [
-        'needle' => ['options' => 'needle'],
+    protected array $messageVariables = [
+        'needle' => 'needle',
     ];
 
-    protected $options = [
-        'needle' => '',
-        'ignoreCase' => false,
-    ];
+    protected readonly ?string $needle;
 
-    protected function testValue($value): void
+    private readonly bool $ignoreCase;
+
+    public function __construct(array $options = [])
+    {
+        $check = array_key_exists('needle', $options);
+
+        if (!$check) {
+            throw new InvalidArgumentException("'needle' is required.");
+        }
+
+        $check = (
+            \is_string($options['needle'])
+            || \is_null($options['needle'])
+            || $options['needle'] instanceof \Stringable
+        );
+
+        if (!$check) {
+            throw new InvalidArgumentException("'needle' must be string or instance of \Stringable");
+        }
+
+        $this->needle = $options['needle'];
+        $this->ignoreCase = isset($options['ignoreCase']) && is_bool($options['ignoreCase']) && $options['ignoreCase'];
+
+        parent::__construct($options);
+    }
+
+    protected function testValue(mixed $value): void
     {
         if (\is_scalar($value)) {
             $haystack = (string) $value;
-            $needle = (string) $this->options['needle'];
+            $needle = (string) $this->needle;
 
-            if ($this->options['ignoreCase']) {
+            if ($this->ignoreCase) {
                 $haystack = strtolower($haystack);
                 $needle = strtolower($needle);
             }
@@ -51,14 +72,5 @@ final class ContainsString extends ExtendedAbstractValidator
         }
 
         $this->error(self::NOT_CONTAINS_STRING);
-    }
-
-    protected function setNeedle($needle)
-    {
-        if (!\is_string($needle) && !\is_null($needle) && !$needle instanceof \Stringable) {
-            throw new InvalidArgumentException("Invalid 'needle': Must be string or instance of \Stringable");
-        }
-
-        $this->options['needle'] = $needle;
     }
 }