+++ tests/Unit/ComparisonTest.php
@@ -11,86 +11,84 @@
 
 declare(strict_types=1);
 
-namespace Alley\Validator;
+namespace Alley\Validator\Tests\Unit;
 
+use Alley\Validator\Comparison;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 final class ComparisonTest extends TestCase
 {
-    /**
-     * @dataProvider dataValidInput
-     */
+    #[DataProvider('dataValidInput')]
     public function testValidInput($value, $options)
     {
         $validator = new Comparison($options);
         $this->assertTrue($validator->isValid($value));
     }
 
-    public function dataValidInput()
+    public static function dataValidInput()
     {
         return [
             '==' => [
                 42,
                 [
                     'operator' => '==',
-                    'compared' => '42',
+                    'target' => '42',
                 ],
             ],
             '===' => [
                 42,
                 [
                     'operator' => '===',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
             ],
             '!=' => [
                 43,
                 [
                     'operator' => '<>',
-                    'compared' => '42',
+                    'target' => '42',
                 ],
             ],
             '!==' => [
                 43,
                 [
                     'operator' => '!==',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
             ],
             '<' => [
                 41,
                 [
                     'operator' => '<',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
             ],
             '>' => [
                 43,
                 [
                     'operator' => '>',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
             ],
             '<=' => [
                 42,
                 [
                     'operator' => '<=',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
             ],
             '>=' => [
                 42,
                 [
                     'operator' => '>=',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
             ],
         ];
     }
 
-    /**
-     * @dataProvider dataInvalidInput
-     */
+    #[DataProvider('dataInvalidInput')]
     public function testInvalidInput($value, $options, $messages)
     {
         $validator = new Comparison($options);
@@ -98,14 +96,14 @@ final class ComparisonTest extends TestCase
         $this->assertSame($messages, $validator->getMessages());
     }
 
-    public function dataInvalidInput()
+    public static function dataInvalidInput()
     {
         return [
             '==' => [
                 43,
                 [
                     'operator' => '==',
-                    'compared' => '42',
+                    'target' => '42',
                 ],
                 ['notEqual' => 'Must be equal to 42 but is 43.'],
             ],
@@ -113,7 +111,7 @@ final class ComparisonTest extends TestCase
                 '42',
                 [
                     'operator' => '===',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
                 ['notIdentical' => 'Must be identical to 42 but is 42.'],
             ],
@@ -121,7 +119,7 @@ final class ComparisonTest extends TestCase
                 42,
                 [
                     'operator' => '<>',
-                    'compared' => '42',
+                    'target' => '42',
                 ],
                 ['isEqual' => 'Must not be equal to 42 but is 42.'],
             ],
@@ -129,7 +127,7 @@ final class ComparisonTest extends TestCase
                 42,
                 [
                     'operator' => '!==',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
                 ['isIdentical' => 'Must not be identical to 42.'],
             ],
@@ -137,7 +135,7 @@ final class ComparisonTest extends TestCase
                 43,
                 [
                     'operator' => '<',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
                 ['notLessThan' => 'Must be less than 42 but is 43.'],
             ],
@@ -145,7 +143,7 @@ final class ComparisonTest extends TestCase
                 41,
                 [
                     'operator' => '>',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
                 ['notGreaterThan' => 'Must be greater than 42 but is 41.'],
             ],
@@ -153,7 +151,7 @@ final class ComparisonTest extends TestCase
                 43,
                 [
                     'operator' => '<=',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
                 ['notLessThanOrEqualTo' => 'Must be less than or equal to 42 but is 43.'],
             ],
@@ -161,7 +159,7 @@ final class ComparisonTest extends TestCase
                 41,
                 [
                     'operator' => '>=',
-                    'compared' => 42,
+                    'target' => 42,
                 ],
                 ['notGreaterThanOrEqualTo' => 'Must be greater than or equal to 42 but is 41.'],
             ],
@@ -172,8 +170,8 @@ final class ComparisonTest extends TestCase
     {
         $operator = 'foo';
 
-        $this->expectExceptionMessageMatches("/^Invalid 'operator': {$operator}\.$/");
+        $this->expectExceptionMessageMatches("/^'operator' must be one of .+?, got {$operator}\.$/");
 
-        new Comparison([ 'operator' => $operator ]);
+        new Comparison(['operator' => $operator, 'target' => 42]);
     }
 }